scripts/requirements.txt
numpy>=1.25.1
Pillow>=10.0.0
pyheif>=0.7.1
labelme>=5.2.1
tqdm>=4.65.0